Bump to newest version in 3.x line
Uncommenting a bunch of stuff that we'll have to deal without for a while. We're
in a weird limbo at this point, where we should really be using Middleman 4, but
we depend on stuff that's not ready for that.

This site is currently in a non-working state, primarily because stuff fails
when we require rails-html-sanitizer.

Also not entirely certain we can actually build and deploy the site at this point.