Rely on browser support for columns
The support is currently good enough, although with many browsers not supporting
the full spec according to https://caniuse.com/#search=column:

> Partial support refers to not supporting the break-before, break-after, break-
> inside properties. WebKit- and Blink-based browsers do have equivalent support
> for the non-standard -webkit-column-break-* properties to accomplish the same
> result (but only the auto and always values). Firefox does not support
> break-*.

We don't use `column-break` so we should be good. Also in the worst case where a
browser doesn't support columns it'll just render a regular list, which is
perfectly fine.