Simple script to update the project
I considered just running script/setup, because the scripts are the same - at
least for now, but it turns out that's not as easy as I'd like it to be.

See
https://stackoverflow.com/questions/38978650/run-a-script-in-the-same-directory-as-the-current-script
for discussions around running another script in the same folder, without
knowing what folder the original script is run from.

In the end, copy/paste was simpler.